code-review-start: Only insert merge_bases row if argument was passed
Don't create a `merge_bases` entry in the database if no merge base
argument was passed to the script. The `get_merge_base` function in
`code-review-database` will default to `origin/master` or `master` when
no row is found in the database. Rely on those defaults instead of
always inserting a row.