code-review-database: Always try to create the database
If the database doesn't exist, we get this error:

    Error: near line 3: no such table: merge_bases

This is because `sqlite3` will always create a new database if it
doesn't exist when running a command against a database. The
newly-created database doesn't have the required table, and we get the
error.

Fix the problem by ensuring the database always exists and is
initialised properly.