code-review-gh-start: Get the merge base from the GitHub API
Haven't tested this yet, but the idea is to use the `gh` CLI to call the
GitHub API and get the list of pulls requests. These are filtered to get
the one open for the current branch, and we can then get its merge base
and start a `code-review` review using that merge base.